resource-server/src/main/java/com/inhabas/api/domain/majorInfo/domain/valueObject/Major.java
package com.inhabas.api.domain.majorInfo.domain.valueObject;

import javax.persistence.Column;
import javax.persistence.Embeddable;
import javax.persistence.Transient;
import java.util.Objects;

@Embeddable
public class Major {

    @Column(name = "major", length = 50, nullable = false)
    private String value;

    @Transient
    private final int MAX_LENGTH = 50;

    public Major() {}

    public Major(String value) {
        if (validate(value))
            this.value = value;
        else
            throw new IllegalArgumentException();
    }

    private boolean validate(Object value) {
        if (Objects.isNull(value)) return false;
        if (!(value instanceof String))  return false;

        String o = (String) value;
        if (o.isBlank()) return false;
        return o.length() < MAX_LENGTH;
    }

    public String getValue() {
        return value;
    }
}
